+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Please add one entry in this file for each change in Yarn's behavior. Use the same format for all entries, including the third-person verb. Make sure you don't add more than one line of text to keep it clean. Thanks!
 
 ## Master
+- Stop `yarn workspaces run` error if command not defined in workspace
 
 - Folders like `.cache` won't be pruned from the `node_modules` after each install.